option 2 to fix grains exercise

+++ exercises/grains/example.jl
@@ -1,7 +1,7 @@
 """Calculate the number of grains on square `square`."""
 function on_square(square)
     check_square_input(square)
-    2 ^ (square - 1)
+    Int128(2) ^ (square - 1)
 end
 
 """Calculate the total number of grains after square `square`."""
